main criteria done

+++ migrations/1676815159469_create-table-albums-&-songs.js
@@ -12,7 +12,7 @@ exports.up = pgm => {
       notNull: true,
     },
     year: {
-      type: 'NUMERIC',
+      type: 'INTEGER',
       notNull: true,
     },
   });
@@ -27,7 +27,7 @@ exports.up = pgm => {
       notNull: true,
     },
     year: {
-      type: 'NUMERIC',
+      type: 'INTEGER',
       notNull: true,
     },
     genre: {
@@ -39,7 +39,7 @@ exports.up = pgm => {
       notNull: true,
     },
     duration: {
-      type: 'NUMERIC',
+      type: 'INTEGER',
     },
     album_id: {
       type: 'VARCHAR(50)',
